Extract type helper into a separate file
